check oscillator (#109)
* feat: add reference to CartPoleCost observation

This commit adds the reference to the CartPoleCost environment
observations. This is needed to train the reference tracking task.

* fix: improve consistency with Han 2020

This commit ensures that the oscillator is consistent with the results
of [Han et al. 2020](https://arxiv.org/abs/2004.14288). It also adds the
ability to set the reset bounds in the reset function and cleans up the
code a bit.

* feat: change info dict

* refactor: cleanup code