build(simzoo): rename `simzoo` package to `stable_gym`
This commit renames the `simzoo` package to `stable-gym`. This was done since the package is no
longer included in the bayesian_learning_control package and in my oppinion the name `stable_gym`
better represents the package content.

BREAKING CHANGE: The package name is changed so the package should be imported as `stable_gym` in
the future.