add ant cost (#184)
* feat: add 'HalfCheetahCost' environment

This commit adds a modified version of the
[HalfCeetah environment](https://gymnasium.farama.org/environments/mujoco/half_cheetah/)
found in the [gymnasium](https://gymnasium.farama.org) package. In this modified
version, the cost was replaced by a reward. This cost is the squared difference
between HalfCheetah's forward velocity and a reference value (error).

* feat: add 'AntCost' environment

This commit adds a modified version of the
[Ant environment](https://gymnasium.farama.org/environments/mujoco/ant/)
found in the [gymnasium](https://gymnasium.farama.org) package. In this modified
version, the cost was replaced by a reward. This cost is the squared difference
between Ant's forward velocity and a reference value (error).

* test: add 'AntCost' tests