feat: add 'MinitaurCost' environment (#222)
This commit adds a modified version of the
[Minitaur environment](https://github.com/bulletphysics/bullet3/tree/master/examples/pybullet/gym/pybullet_data/quadruped)
found in the [bullet3 package](https://pybullet.org/wordpress/) package.
In this modified version, the objective was changed to a
velocity-tracking task. To do this, the reward is replaced with a cost.
This cost is the squared difference between the Minitaur's forward
velocity and a reference value (error).